Ouroboros.Consensus.Example.Block

Eras

type ExampleEras c = '[ShelleyBlock (ShelleyEra c), ShelleyBlock (ExampleEra c)]

The eras in the Example blockchain.

We parameterise over the crypto: c.

module Ouroboros.Consensus.Example.Eras

type ShelleyBasedExampleEras c = '[ShelleyEra c, ExampleEra c]

Block

type ExampleBlock c = HardForkBlock (ExampleEras c)

The Example block.

Thanks to the pattern synonyms, you can treat this as a sum type with constructors BlockShelley and BlockExample.

f :: ExampleBlock c -> _
f (BlockShelley s) = _
f (BlockExample a) = _

data HardForkBlock (xs ∷ [Type]) where

Bundled Patterns

pattern BlockShelley ∷ ShelleyBlock (ShelleyEra c) → ExampleBlock c

pattern BlockExample ∷ ShelleyBlock (ExampleEra c) → ExampleBlock c

Headers

type ExampleHeader c = Header (ExampleBlock c)

The Example header.

data family Header blk

Generalised transactions

type ExampleApplyTxErr c = HardForkApplyTxErr (ExampleEras c)

An error resulting from applying a ExampleGenTx to the ledger.

Thanks to the pattern synonyms, you can treat this as a sum type with constructors ApplyTxErrShelley, ApplyTxErrExample, and ApplyTxErrWrongEra.

toText :: ExampleApplyTxErr c -> Text
toText (ApplyTxErrShelley s) = shelleyApplyTxErrToText s
tlText (ApplyTxErrExample a) = exampleApplyTxErrToText a
toText (ApplyTxErrWrongEra eraMismatch) =
  "Transaction from the " <> otherEraName eraMismatch <>
  " era applied to a ledger from the " <>
  ledgerEraName eraMismatch <> " era"

type ExampleGenTx c = GenTx (ExampleBlock c)

The Example transaction.

type ExampleGenTxId c = GenTxId (ExampleBlock c)

The ID of an Example transaction.

data family GenTx blk

Generalized transaction

The mempool (and, accordingly, blocks) consist of "generalized transactions"; this could be "proper" transactions (transferring funds) but also other kinds of things such as update proposals, delegations, etc.

data HardForkApplyTxErr (xs ∷ [Type]) where

Bundled Patterns

pattern ApplyTxErrShelley ∷ ApplyTxErr (ShelleyBlock (ShelleyEra c)) → ExampleApplyTxErr c

pattern ApplyTxErrExample ∷ ApplyTxErr (ShelleyBlock (ExampleEra c)) → ExampleApplyTxErr c

pattern ApplyTxErrWrongEra ∷ EraMismatch → ExampleApplyTxErr c

data family TxId tx

A generalized transaction, GenTx, identifier.

LedgerError

type ExampleLedgerError c = HardForkLedgerError (ExampleEras c)

An error resulting from applying a ExampleBlock to the ledger.

Thanks to the pattern synonyms, you can treat this as a sum type with constructors LedgerErrorShelley, LedgerErrorExample, and LedgerErrorWrongEra.

toText :: ExampleLedgerError c -> Text
toText (LedgerErrorShelley s) = shelleyLedgerErrorToText s
toText (LedgerErrorExample a) = allegraLedgerErrorToText a
toText (LedgerErrorWrongEra eraMismatch) =
  "Block from the " <> otherEraName eraMismatch <>
  " era applied to a ledger from the " <>
  ledgerEraName eraMismatch <> " era"

data HardForkLedgerError (xs ∷ [Type]) where

Bundled Patterns

pattern LedgerErrorShelley ∷ LedgerError (ShelleyBlock (ShelleyEra c)) → ExampleLedgerError c

pattern LedgerErrorExample ∷ LedgerError (ShelleyBlock (ExampleEra c)) → ExampleLedgerError c

pattern LedgerErrorWrongEra ∷ EraMismatch → ExampleLedgerError c

OtherEnvelopeError

type ExampleOtherHeaderEnvelopeError c = HardForkEnvelopeErr (ExampleEras c)

An error resulting from validating a ExampleHeader.

data HardForkEnvelopeErr (xs ∷ [Type]) where

Bundled Patterns

pattern OtherHeaderEnvelopeErrorShelley ∷ OtherHeaderEnvelopeError (ShelleyBlock (ShelleyEra c)) → ExampleOtherHeaderEnvelopeError c

pattern OtherHeaderEnvelopeErrorExample ∷ OtherHeaderEnvelopeError (ShelleyBlock (ExampleEra c)) → ExampleOtherHeaderEnvelopeError c

pattern OtherHeaderEnvelopeErrorWrongEra ∷ EraMismatch → ExampleOtherHeaderEnvelopeError c

TipInfo

type ExampleTipInfo c = OneEraTipInfo (ExampleEras c)

The TipInfo of the Example chain.

data OneEraTipInfo (xs ∷ [Type]) where

Bundled Patterns

pattern TipInfoShelley ∷ TipInfo (ShelleyBlock (ShelleyEra c)) → ExampleTipInfo c

pattern TipInfoExample ∷ TipInfo (ShelleyBlock (ExampleEra c)) → ExampleTipInfo c

Query

data Either a b where

The Either type represents values with two possibilities: a value of type Either a b is either Left a or Right b.

The Either type is sometimes used to represent a value which is either correct or an error; by convention, the Left constructor is used to hold an error value and the Right constructor is used to hold a correct value (mnemonic: "right" also means "correct").

Examples

Expand

The type Either String Int is the type of values which can be either a String or an Int. The Left constructor can be used only on Strings, and the Right constructor can be used only on Ints:

>>> let s = Left "foo" :: Either String Int
>>> s
Left "foo"
>>> let n = Right 3 :: Either String Int
>>> n
Right 3
>>> :type s
s :: Either String Int
>>> :type n
n :: Either String Int

The fmap from our Functor instance will ignore Left values, but will apply the supplied function to values contained in a Right:

>>> let s = Left "foo" :: Either String Int
>>> let n = Right 3 :: Either String Int
>>> fmap (*2) s
Left "foo"
>>> fmap (*2) n
Right 6

The Monad instance for Either allows us to chain together multiple actions which may fail, and fail overall if any of the individual steps failed. First we'll write a function that can either parse an Int from a Char, or fail.

>>> import Data.Char ( digitToInt, isDigit )
>>> :{
    let parseEither :: Char -> Either String Int
        parseEither c
          | isDigit c = Right (digitToInt c)
          | otherwise = Left "parse error"
>>> :}

The following should work, since both '1' and '2' can be parsed as Ints.

>>> :{
    let parseMultiple :: Either String Int
        parseMultiple = do
          x <- parseEither '1'
          y <- parseEither '2'
          return (x + y)
>>> :}

>>> parseMultiple
Right 3

But the following should fail overall, since the first operation where we attempt to parse 'm' as an Int will fail:

>>> :{
    let parseMultiple :: Either String Int
        parseMultiple = do
          x <- parseEither 'm'
          y <- parseEither '2'
          return (x + y)
>>> :}

>>> parseMultiple
Left "parse error"

Bundled Patterns

pattern QueryResultSuccess ∷ result → ExampleQueryResult c result

pattern QueryResultEraMismatch ∷ EraMismatch → ExampleQueryResult c result
A query from a different era than the ledger's era was sent.

type ExampleQuery c = Query (ExampleBlock c)

The Query of Example chain.

type ExampleQueryResult c = HardForkQueryResult (ExampleEras c)

The result of a ExampleQuery

Thanks to the pattern synonyms, you can treat this as a sum type with constructors QueryResultSuccess and QueryResultEraMismatch.

data family Query blk ∷ Type → Type

Different queries supported by the ledger, indexed by the result type.

CodecConfig

data family CodecConfig blk

Static configuration required for serialisation and deserialisation of types pertaining to this type of block.

Data family instead of type family to get better type inference.

type ExampleCodecConfig c = CodecConfig (ExampleBlock c)

The CodecConfig for ExampleBlock.

Thanks to the pattern synonyms, you can treat this as the product of the Shelley, Example ... CodecConfigs.

BlockConfig

data family BlockConfig blk

Static configuration required to work with this type of blocks

type ExampleBlockConfig c = BlockConfig (ExampleBlock c)

The BlockConfig for ExampleBlock.

Thanks to the pattern synonyms, you can treat this as the product of the Shelley, Example, ... BlockConfigs.

StorageConfig

type ExampleStorageConfig c = StorageConfig (ExampleBlock c)

The StorageConfig for ExampleBlock.

Thanks to the pattern synonyms, you can treat this as the product of the Shelley, Example, ... StorageConfigs.

data family StorageConfig blk

Config needed for the NodeInitStorage class. Defined here to avoid circular dependencies.

ConsensusConfig

data family ConsensusConfig p

Static configuration required to run the consensus protocol

Every method in the ConsensusProtocol class takes the consensus configuration as a parameter, so having this as a data family rather than a type family resolves most ambiguity.

Defined out of the class so that protocols can define this type without having to define the entire protocol at the same time (or indeed in the same module).

type ExampleConsensusConfig c = ConsensusConfig (HardForkProtocol (ExampleEras c))

The ConsensusConfig for ExampleBlock.

Thanks to the pattern synonyms, you can treat this as the product of the Shelley, Example, ... PartialConsensusConfigs.

NOTE: not ConsensusConfig, but PartialConsensusConfig.

LedgerConfig

type ExampleLedgerConfig c = HardForkLedgerConfig (ExampleEras c)

The LedgerConfig for ExampleBlock.

Thanks to the pattern synonyms, you can treat this as the product of the Shelley, Example, ... PartialLedgerConfigs.

NOTE: not LedgerConfig, but PartialLedgerConfig.

data HardForkLedgerConfig (xs ∷ [Type]) where

Bundled Patterns

pattern ExampleLedgerConfig ∷ PartialLedgerConfig (ShelleyBlock (ShelleyEra c)) → PartialLedgerConfig (ShelleyBlock (ExampleEra c)) → ExampleLedgerConfig c

LedgerState

type ExampleLedgerState c = LedgerState (ExampleBlock c)

The LedgerState for ExampleBlock.

NOTE: the ExampleLedgerState contains more than just the current era's LedgerState. We don't give access to those internal details through the pattern synonyms. This is also the reason the pattern synonyms are not bidirectional.

data family LedgerState blk

Ledger state associated with a block

ChainDepState

type ExampleChainDepState c = HardForkChainDepState (ExampleEras c)

The ChainDepState for ExampleBlock.

NOTE: the ExampleChainDepState contains more than just the current era's ChainDepState. We don't give access to those internal details through the pattern synonyms. This is also the reason the pattern synonyms are not bidirectional.

data HardForkState (f ∷ Type → Type) (xs ∷ [Type]) where

Generic hard fork state

This is used both for the consensus state and the ledger state.

Bundled Patterns

pattern ChainDepStateShelley ∷ ChainDepState (BlockProtocol (ShelleyBlock (ShelleyEra c))) → ExampleChainDepState c

pattern ChainDepStateExample ∷ ChainDepState (BlockProtocol (ShelleyBlock (ExampleEra c))) → ExampleChainDepState c

EraMismatch

data EraMismatch

Extra info for errors caused by applying a block, header, transaction, or query from one era to a ledger from a different era.

Constructors

EraMismatch

Fields

ledgerEraName ∷ !Text

Name of the era of the ledger (Byron or Shelley).

otherEraName ∷ !Text

Era of the block, header, transaction, or query.