Ouroboros.Consensus.Example.Eras

Eras based on the Shelley ledger

data ExampleEra c

Eras instantiated with standard crypto

type StandardExample = ExampleEra StandardCrypto

The Example era with standard crypto

Re-exports

class (ShelleyBasedEra era, State (EraRule "PPUP" era) ~ PPUPState era, Default (State (EraRule "PPUP" era)), HasField "_maxBHSize" (PParams era) Natural, HasField "_maxTxSize" (PParams era) Natural, HasField "_a0" (PParams era) Rational, HasField "_nOpt" (PParams era) Natural, HasField "_rho" (PParams era) UnitInterval, HasField "_tau" (PParams era) UnitInterval, HasField "_protocolVersion" (PParamsDelta era) (StrictMaybe ProtVer), AdditionalGenesisConfig era ~ (), FromCBOR (PParams era), FromCBOR (PParamsDelta era), ToCBOR (PParams era)) ⇒ ShelleyBasedEra era where

The ledger already defines ShelleyBasedEra as the top-level constraint on an era, however, consensus often needs some more functionality than the ledger currently provides.

Either the functionality shouldn't or can't live in the ledger, in which case it can be part and remain part of ShelleyBasedEra. Or, the functionality should live in the ledger, but hasn't yet been added to the ledger, or it hasn't yet been propagated to this repository, in which case it can be added to this class until that is the case.

By having the same name as the class defined in ledger, we can, if this class becomes redundant, switch to the ledger-defined one without having to update all the code using it. We can just export the right one from this module.

TODO Currently we include some constraints on the update state which are needed to determine the hard fork point. In the future this should be replaced with an appropriate API - see https://github.com/input-output-hk/ouroboros-network/issues/2890

TODO Currently we include the constraint SL.AdditionalGenesisConfig era ~ (). When we fork to Alonzo we will need additional genesis config information.

Methods

shelleyBasedEraName ∷ proxy era → Text

Return the name of the Shelley-based era, e.g., Shelley, Allegra, etc.

type EraCrypto era = Crypto era

The Crypto type family conflicts with the Crypto class. To avoid having to import one or both of them qualified, define EraCrypto as an alias of the former: /return the crypto used by this era/.

data ShelleyEra c

type StandardShelley = ShelleyEra StandardCrypto

The Shelley era with standard crypto

data StandardCrypto

Orphan instances

PraosCrypto c ⇒ ShelleyBasedEra (ExampleEra c)

Instance details

Methods

shelleyBasedEraName ∷ proxy (ExampleEra c) → Text